フロントエンド：コンポーネント化修正
