+++ apps/frontend/src/features/auth/types/auth.ts
@@ -1,0 +1,5 @@
+export type Auth = {
+  signUp: (email: string, password: string) => Promise<string>;
+  signIn: (email: string, password: string) => Promise<string>;
+  signOut: () => Promise<string>;
+};
